requirements.txt
# Shadow Detection Preprocessing Dependencies
# Reference: Vicente et al., "Leave-One-Out Kernel Optimization for Shadow Detection", ICCV 2015

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Image processing
scikit-image>=0.19.0
Pillow>=8.0.0

# Machine learning (for later stages)
scikit-learn>=1.0.0

# Visualization (optional)
matplotlib>=3.5.0

# MRF optimization (Section 4 of the paper)
PyMaxflow>=1.3.0

# Note: For additional functionality:
# - opencv-python (for advanced image processing)

